Rename and fix custom sprite source test
