+++ tests/src/main/resources/META-INF/mods.toml
@@ -166,7 +166,7 @@ modId="fishing_bobber_event_test"
 [[mods]]
 modId="custom_respawn_test"
 [[mods]]
-modId="custom_tas_test"
+modId="custom_sprite_source_test"
 [[mods]]
 modId="dimension_settings_test"
 [[mods]]
