+++ tests/src/main/resources/assets/custom_sprite_source_test/models/item/test_item.json
@@ -1,6 +1,6 @@
 {
   "parent": "minecraft:item/handheld",
   "textures": {
-    "layer0": "custom_tas_test:item/test_item"
+    "layer0": "custom_sprite_source_test:test_item"
   }
 }
