+++ tests/src/main/resources/assets/custom_sprite_source_test/textures/test_item.png
