--- tests/src/main/resources/assets/custom_tas_test/textures/item/test_item.png.mcmeta
@@ -1,5 +1,0 @@
-{
-  "forge": {
-    "loader": "custom_tas_test:tas_loader"
-  }
-}
